requirements.txt
langchain==0.3.27
langchain-core==0.3.79
langchain-community==0.3.31
langgraph==0.6.10
langgraph-checkpoint==2.1.2
dashscope==1.24.6
mcp==1.17.0
pyyaml==6.0.3
rich==14.2.0
